pyproject.toml
[tool.poetry]
name = "wit-core"
version = "0.1.0"
description = "Response processor for wit.ai"
authors = ["Lucas Oliveira <ls.oliveiras.santos@gmail.com>"]
repository = "https://github.com/LucasOliveiraS/wit-core"
documentation = "https://github.com/LucasOliveiraS/wit-core"
classifiers = ["Programming Language :: Python :: 3", "License :: OSI Approved :: MIT License","Operating System :: OS Independent",]
keywords = ["bot", "bots", "botkit", "wit", "wit-ai", "wit-core", "conversational-ai", "chatbot", "chatbot-framework", "bot-framework",]
readme = "README.md"
license = "MIT"
include = ["LICENSE", "README.md"]

[tool.poetry.dependencies]
python = ">3.6.1"
wheel = "^0.36.2"
pampy = "^0.3.0"
PyYAML = "^5.4.1"
click = "^7.1.2"
wit = "^6.0.0"
python-dotenv = "^0.17.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.3"
mypy = "^0.812"
pytest-mock = "^3.5.1"
flake8 = "^3.9.0"
pre-commit = "^2.12.0"
autopep8 = "^1.5.6"

[tool.poetry.scripts]
wit-core = 'wit_core.cli:cli'

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"